Implemnetation/requirements.txt
numpy>=1.26
pandas>=2.0
scipy>=1.11
pywavelets>=1.5.0
scikit-learn>=1.4
xgboost>=2.0
pytorch-tabnet>=4.1
torch>=2.2
tqdm>=4.66
matplotlib>=3.8
seaborn>=0.13
shap>=0.45
lime>=0.2.0.1
umap-learn>=0.5.6
joblib>=1.4
pyyaml>=6.0
